Can now do its own MX lookup.
When the REMOTESYS of a outgoing queue is set to "--MX--", then the SMTP
server now does MX lookups to determine the server(s) to send each message
to.

Just sending a single message over a existing SMTP connection was
extracted into the new subroutine SMTP_SEND_MESSAGE.  This new subroutine
should now be used by the existing code that sends to a fixed host in the
SMTP_SEND module.  This has not yet been done.  There is now duplicate
sending code in the SMTP_SEND and SMTP_SEND_MESSAGE modules.