Added work-around for creating a new queue entry.
In some cases, the message file of a queue entry is apparently still hung
open, even though the control file can be opened with read+write access.
That is not supposed to happen.  The cause of this was not found, and
should still be investigated and fixed.

However, the logic to create a new queue entry was made more robust.  The
previous logic would assume the sequence number was found when the control
file could be opened with read+write access, and it's length was 0.  That
should have been good enough.  Due to the problem described above, this
led to a error when trying to create the message file.  Now the next
sequence number is tried if there is any error in opening any of the three
queue entry files, not just the control file.  This seems to work.

Also deleted lots of unused stuff from the EMAIL.MSG file.  This was
originally cloned from the message file of the old STUFF library, which
contained a lot more than just email manipulation.  The other non-email
stuff was not cleaned out, and has now been removed.